requirements.txt
astroid==1.6.1
autopep8==1.3.5
backcall==0.1.0
cached-property==1.4.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==1.11
django-adminlte2==0.3.0
django-extensions==2.0.7
django-money==0.12.3
django-pyodbc-azure==1.11.0.0
django-python3-ldap==0.11.1
django-reset-migrations==0.3.1
django-url-filter==0.3.5
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
enum-compat==0.0.2
et-xmlfile==1.0.1
idna==2.7
isort==4.3.4
itypes==1.1.0
jdcal==1.3
jedi==0.12.0
Jinja2==2.10
lazy-object-proxy==1.3.1
ldap3==2.3
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
openpyxl==2.5.0
pandas==0.22.0
parso==0.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
py-moneyed==0.7.0
pyasn1==0.3.2
pycodestyle==2.4.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==1.8.2
pyodbc==4.0.21
python-dateutil==2.6.1
pytz==2017.3
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
